Fix crash when no samples are recorded
BZ: 118157

Fix a NULL pointer reference the occurs because the
vmlinux_maps table is assigned only if sample events
are recorded in the perf.data file.

Depending on the process/thread selected, the activity
level of the process/thread, and the duration of the
sleep, it appears valid for no sample events to be
recorded. If no samples are recorded, extend the sleep
interval.

NOTE:  Pre-exising problem displaying sample count on record
persists. This does not, however, affect the accuracy of reports.

Change-Id: I27525e038ee7c7041e72a8164564439f9c24bbe9
Reviewed-on: http://android.intel.com:8080/118866
Reviewed-by: Smith, Geoff <geoff.smith@intel.com>
Reviewed-by: Yang, Dong <dong.yang@intel.com>
Reviewed-by: Wang, LinX Z <linx.z.wang@intel.com>
Reviewed-by: Balakrishnan, Suresh <suresh.balakrishnan@intel.com>
Tested-by: Ranostay, Matt <matt.ranostay@intel.com>
Reviewed-by: Noziska, Patrick J <patrick.j.noziska@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>